web: remove jQuery and Font Awesome CDN, embed all assets locally
- Extract all JS from index_html.h into new index_js.h (served as /index.js)
- Replace jQuery (~90KB PROGMEM) with minimal vanilla polyfill
- Embed 12 SVG icons as CSS mask-image data URIs, removing FA CDN dependency
- Remove FA detection fallback code from JS
- Bump ETags: app-v2 (JS), style-v2 (CSS)

Changelog: Dashboard loads faster and works fully offline — removed jQuery
and Font Awesome CDN; icons and JS now embedded in firmware
