+++ .instructions.md
@@ -62,6 +62,8 @@ After the workflow output:
 - ✅ You have full control over commit message and content
 - ✅ Files are staged but not committed
 - ✅ Review everything before committing
+- ✅ **CHANGELOG.md must always be updated and staged before any commit** — no exceptions
+- ✅ The agent must never run `git commit` without explicit user instruction to do so
 
 ### Reverting Staged Changes
